[Refactor] create string.c and move some functions into string.c
