+++ lab1/shell.c
@@ -6,28 +6,6 @@
 #include "shell.h"
 #include "uart.h"
 
-/* itoa impl reference: */
-/* https://stackoverflow.com/questions/3982320/convert-integer-to-string-without-access-to-libraries */
-char* itoa(int val, int base){
-  static char buf[32] = {0};
-  int i = 30;
-  for(; val && i ; --i, val /= base)
-    buf[i] = "0123456789abcdef"[val % base];
-  return &buf[i+1];
-}
-
-void hello() {
-  uart_puts("Hello World!\r\n");
-}
-
-int sstrcmp(const char *s1, const char *s2) {
-  while (*s1 && (*s1 == *s2)) {
-    s1++;
-    s2++;
-  }
-  return *s1 - *s2;
-}
-
 int getcmd(char *buf, int nbuf) {
   uart_puts("# ");
