+++ lab1/string.h
@@ -1,0 +1,4 @@
+#pragma once
+
+char* itoa(int val, int base);
+int sstrcmp(const char *s1, const char *s2);
